requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
beautifulsoup4==4.14.2
billiard==4.2.4
bleach==6.2.0
boto3==1.38.46
botocore==1.38.46
bs4==0.0.2
celery==5.6.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-modeltranslation==0.19.16
django-rest-framework==0.1.0
django-storages==1.14.6
django-summernote==0.8.20.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.3.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.12.0
jmespath==1.0.1
kombu==5.6.1
openai==2.8.1
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==7.1.0
requests==2.32.5
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
whitenoise==6.9.0
